remove org.json:json in favour of AOSP version
`org.json:json` has a non-standard license. This commit drops it from
the POM and replaces it with the Apache 2 licensed clean room
reimplementation done by Google as part of the Android Open Source
Project.

See
https://android.googlesource.com/platform/libcore/+/ics-plus-aosp/json/
for original code.

This doesn't build - a future commit will patch the necessary places in
the CC to adapt call sites for some incompatible API changes in Google's
reimplementation.